Update the Standard endpoint interface and simplify it
