+++ edenai_apis/apis/google/google_api.py
@@ -16,7 +16,7 @@ from edenai_apis.apis.google.google_multimodal_api import GoogleMultimodalApi
 from edenai_apis.features import ProviderInterface
 from edenai_apis.loaders.data_loader import ProviderDataEnum
 from edenai_apis.loaders.loaders import load_provider
-from edenai_apis.llmengine.llm_engine import LLMEngine, StdLLMEngine
+from edenai_apis.llmengine.llm_engine import LLMEngine
 from edenai_apis.features.llm.llm_interface import LlmInterface
 
 
@@ -53,11 +53,6 @@ class GoogleApi(
             "translate": translate.TranslationServiceClient(),
             "llm_client": LLMEngine(
                 provider_name="gemini",
-                provider_config={
-                    "api_key": self.api_settings.get("genai_api_key"),
-                },
-            ),
-            "std_llm_client": StdLLMEngine(
                 provider_config={
                     "api_key": self.api_settings.get("genai_api_key"),
                 },
