+++ edenai_apis/apis/mistral/mistral_api.py
@@ -18,7 +18,7 @@ from edenai_apis.loaders.data_loader import ProviderDataEnum
 from edenai_apis.loaders.loaders import load_provider
 from edenai_apis.utils.exception import ProviderException
 from edenai_apis.utils.types import ResponseType
-from edenai_apis.llmengine.llm_engine import LLMEngine, StdLLMEngine
+from edenai_apis.llmengine.llm_engine import LLMEngine
 from edenai_apis.features.llm.llm_interface import LlmInterface
 from edenai_apis.features.llm.chat.chat_dataclass import ChatDataClass
 
@@ -36,8 +36,6 @@ class MistralApi(ProviderInterface, TextInterface, LlmInterface):
         self.llm_client = LLMEngine(
             provider_name=self.provider_name, provider_config={"api_key": self.api_key}
         )
-
-        self.std_llm_client = StdLLMEngine(provider_config={"api_key": self.api_key})
 
     def text__generation(
         self, text: str, temperature: float, max_tokens: int, model: str, **kwargs
@@ -193,7 +191,7 @@ class MistralApi(ProviderInterface, TextInterface, LlmInterface):
         # Optional parameters
         **kwargs,
     ) -> ChatDataClass:
-        response = self.std_llm_client.completion(
+        response = self.llm_client.completion(
             messages=messages,
             model=model,
             timeout=timeout,
