+++ edenai_apis/llmengine/__init__.py
@@ -1,2 +1,1 @@
 from .llm_engine import LLMEngine
-from .llm_engine import StdLLMEngine
