+++ edenai_apis/llmengine/llm_engine.py
@@ -105,7 +105,6 @@ class LLMEngine:
     def _execute_completion(self, params: Dict, response_class: Type, **kwargs):
         try:
             params.pop("moderate_content", None)
-            print(f"======================> completion call_params: {params}")
             response = self.completion_client.completion(**params, **kwargs)
             response = ResponseModel.model_validate(response)
             result = json.loads(response.choices[0].message.content)
@@ -151,7 +150,6 @@ class LLMEngine:
                 tools=available_tools
             )
             call_params["tool_choice"] = tool_choice
-        print(f"======================> chat call_params: {call_params}")
         response = self.completion_client.completion(**call_params, **kwargs)
         if stream is False:
             response = ResponseModel.model_validate(response)
@@ -838,157 +836,7 @@ class LLMEngine:
 
             completion_params = completion_params | kwargs
             call_params = self._prepare_args(**completion_params)
-            print(f"======================> std chat call_params: {call_params}")
             response = self.completion_client.completion(**call_params, **kwargs)
-            response = ResponseModel.model_validate(response)
-            return response
-        except Exception as ex:
-            raise ex
-
-
-class StdLLMEngine(LLMEngine):
-
-    PROVIDER_MAPPING = {
-        "vertex_ai": "google",
-        "gemini": "google",
-        "bedrock_converse": "bedrock",
-        "amazon": "bedrock",
-    }
-
-    def __init__(
-        self,
-        provider_config: dict = {},
-        **kwargs,
-    ):
-        super().__init__(
-            model=None,
-            client_name="litellm",
-            application_name="std_chat",
-            provider_config=provider_config,
-            provider_name=None,
-            **kwargs,
-        )
-
-    @staticmethod
-    def map_provider(provider_name: str) -> str:
-        if provider_name is None:
-            return None
-        # Try to regex match the keys of PROVIDER_MAPPING and provider_name. The first one ot match wins
-        for key in StdLLMEngine.PROVIDER_MAPPING.keys():
-            if re.match(key, provider_name, re.RegexFlag.IGNORECASE):
-                return StdLLMEngine.PROVIDER_MAPPING[key]
-        return provider_name
-
-    @moderate_std
-    def completion(
-        self,
-        messages: List = [],
-        model: Optional[str] = None,
-        # Optional OpenAI params: see https://platform.openai.com/docs/api-reference/chat/create
-        timeout: Optional[Union[float, str, httpx.Timeout]] = None,
-        temperature: Optional[float] = None,
-        top_p: Optional[float] = None,
-        n: Optional[int] = None,
-        stream: Optional[bool] = None,
-        stream_options: Optional[dict] = None,
-        stop: Optional[str] = None,
-        stop_sequences: Optional[any] = None,
-        max_tokens: Optional[int] = None,
-        presence_penalty: Optional[float] = None,
-        frequency_penalty: Optional[float] = None,
-        logit_bias: Optional[dict] = None,
-        # openai v1.0+ new params
-        response_format: Optional[
-            Union[dict, Type[BaseModel]]
-        ] = None,  # Structured outputs
-        seed: Optional[int] = None,
-        tools: Optional[List] = None,
-        tool_choice: Optional[Union[str, dict]] = None,
-        logprobs: Optional[bool] = None,
-        top_logprobs: Optional[int] = None,
-        parallel_tool_calls: Optional[bool] = None,
-        deployment_id=None,
-        extra_headers: Optional[dict] = None,
-        # soon to be deprecated params by OpenAI -> This should be replaced by tools
-        functions: Optional[List] = None,
-        function_call: Optional[str] = None,
-        base_url: Optional[str] = None,
-        api_version: Optional[str] = None,
-        api_key: Optional[str] = None,
-        model_list: Optional[list] = None,  # pass in a list of api_base,keys, etc.
-        drop_invalid_params: bool = True,  # If true, all the invalid parameters will be ignored (dropped) before sending to the model
-        user: str | None = None,
-        # Optional parameters
-        **kwargs,
-    ):
-        kwargs.pop("moderate_content", None)
-        # if "provider" in kwargs:
-        #     # Verify if the provider is gemini
-        #     provider_name = kwargs.pop("provider", None)
-        #     is_gemini = provider_name == "gemini"
-        #     provider_name = StdLLMEngine.map_provider(provider_name)
-        #     if provider_name == "google" and not is_gemini:
-        #         api_settings, location = load_provider(
-        #             ProviderDataEnum.KEY,
-        #             provider_name=provider_name,
-        #             location=True,
-        #             api_keys=api_key,
-        #         )
-        #         self.project_id = api_settings["project_id"]
-        #         os.environ["GOOGLE_APPLICATION_CREDENTIALS"] = location
-        #     elif is_gemini:
-        #         api_settings = load_provider(
-        #             ProviderDataEnum.KEY, provider_name=provider_name, api_keys=api_key
-        #         )
-        #         api_key = api_settings["genai_api_key"]
-        #     else:
-        #         api_settings = load_provider(
-        #             ProviderDataEnum.KEY, provider_name=provider_name, api_keys=api_key
-        #         )
-        #         api_key = api_settings["api_key"]
-        try:
-            completion_params = {
-                "messages": messages,
-                "model": model,
-                "timeout": timeout,
-                "temperature": temperature,
-                "top_p": top_p,
-                "n": n,
-                "stream": stream,
-                "stream_options": stream_options,
-                "stop": stop,
-                "stop_sequences": stop_sequences,
-                "max_tokens": max_tokens,
-                "presence_penalty": presence_penalty,
-                "frequency_penalty": frequency_penalty,
-                "logit_bias": logit_bias,
-                "response_format": response_format,
-                "seed": seed,
-                "tools": tools,
-                "tool_choice": tool_choice,
-                "logprobs": logprobs,
-                "top_logprobs": top_logprobs,
-                "parallel_tool_calls": parallel_tool_calls,
-                "deployment_id": deployment_id,
-                "extra_headers": extra_headers,
-                "functions": functions,
-                "function_call": function_call,
-                "base_url": base_url,
-                "api_version": api_version,
-                "api_key": self.provider_config["api_key"],
-                "model_list": model_list,
-                "drop_invalid_params": drop_invalid_params,
-                "user": user,
-                **kwargs,
-            }
-            response = self._execute_completion(completion_params)
-            return response
-        except Exception as ex:
-            raise ex
-
-    def _execute_completion(self, params: Dict, **kwargs):
-        try:
-            response = self.completion_client.completion(**params, **kwargs)
             response = ResponseModel.model_validate(response)
             return response
         except Exception as ex:
